requirements_rtx5090.txt
torch
torchvision
flash-attn>=2.7.0
timm>=1.0.14
albumentations>=2.0.4
onnx>=1.14.02
onnxruntime>=1.15.1
pycocotools>=2.0.7
PyYAML>=6.0.1
scipy>=1.13.0
onnxslim>=0.1.31
gradio>=4.44.1
opencv-python-headless>=4.9.0.80
psutil>=5.9.8
py-cpuinfo>=9.0.0
huggingface-hub>=0.23.2
safetensors>=0.4.3
numpy>=1.26.4,<2.0.0
supervision>=0.22.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.64.0
ultralytics
transformers
seaborn
thop>=0.1.0